% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_functions.R
\name{physeq.convert.rel}
\alias{physeq.convert.rel}
\title{Convert counts to relative abundance.}
\usage{
physeq.convert.rel(physeq, warning.in = TRUE)
}
\arguments{
\item{physeq}{input phyloseq object}

\item{warning.in}{\code{TRUE} or \code{FALSE}. If total relative abundance for a sample does not equal 1, should an error be issued?}
}
\value{
Returns both a phyloseq object and the OTU table as a dataframe where counts are replaced by relative abundance.
}
\description{
Converts taxa counts stored in phyloseq object into taxa relative abundance
}
\examples{

}
