% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_functions.R
\name{physeq.create}
\alias{physeq.create}
\title{Create phyloseq object}
\usage{
physeq.create(
  asvdf,
  taxdf,
  mapdf,
  ID.col,
  export.asvsp = TRUE,
  export.wd = get_wd()
)
}
\arguments{
\item{asvdf}{Dataframe containing ASV sequence count data for each sample. Rows are sequences and column are samples. Rownames must be set to sequence ID..}

\item{taxdf}{Dataframe contaningg taxonomic classification of each ASV.  Rows are sequences and columns are samples. Rownames must be set to sequence ID.}

\item{mapdf}{Dataframe containing sample meta data (e.g clinical variables). Rows are samples and columns are metadata variables (e.g age, sex, clinical variables, etc.)}

\item{ID.col}{#String. Name of the column in mapdf that contains sample IDs. (E.g \code{"sample.ids"})}

\item{export.asvsp}{#\code{TRUE} or \code{FALSE} whether to export a dataframe that contains ASV sequences and their corresponding species ID (spID) auto-generated by phyloseq}

\item{export.wd}{#Directory path as a string (e.g \code{"~/output"}. Location of directory where dataframe will be exported if export.asvp = TRUE.}
}
\value{
Phyloseq object.
}
\description{
Function to create phyloseq obejct, with additional quality control measures to ensure input dataframes are correctly aligned for phyloseq.
}
\examples{
dat <- physeq.create(asvdf = asv, taxdf = taxanomy, mapdf = metadata, ID.col = "sample_ID", export.asvsp = FALSE)

}
