% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_functions.R
\name{physeq.mra.prevalence}
\alias{physeq.mra.prevalence}
\title{Relative abundance Summary and Conversion}
\usage{
physeq.mra.prevalence(physeq, type, vis.export = TRUE, export_prefix.in)
}
\arguments{
\item{physeq}{Input phyloseq object}

\item{type}{One of \code{"counts"} or \code{"relabund"}. Describes whether input phyloseq object currently contains raw counts or relative abundance.}

\item{vis.export}{\code{TRUE} or \code{FALSE}. If \code{"TRUE"}, the function save plots to the working directory describing distribution of taxa relative abundance and prevalence.}

\item{export_prefix.in}{Provide a prefix (string) for summary plots if vis.export = \code{TRUE}}
}
\value{
List containing [1] mean relative abundance and prevalence for each taxa (dataframe) [2] Relative abundance OTU table (dataframe) [3] Relative abundance phyloseq object.
If vis.export = TRUE Will also export summary plots of mean relative abundance and prevalence distributions to working directory
}
\description{
A comprehensive relative abundance function for phyloseq objects. Includes output of relative abundance summary statistics (mean relative abundance and prevalence), both as a dataframe and with an added option for plotting. The function accepts input of count or relative abundance data.
}
\details{
If count data is input, it will be first converted to relative abundance. The relative abundance phyloseq object and OTU table will also be returned.
}
\examples{
physeq.mra.prevalence(physeq = dat, type = "counts", vis.export = TRUE, export_prefix.in = "test_counts") #with export
physeq.mra.prevalence(physeq = dat_rel, type = "relabund") #input phyloseq object with relative abundance
}
