% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_functions.R
\name{sp_taxa_ID}
\alias{sp_taxa_ID}
\title{Species ID taxonomoc classification}
\usage{
sp_taxa_ID(physeq, taxa.col = NA, remove = FALSE, remove.delimiter = "_")
}
\arguments{
\item{physeq}{Phyloseq object}

\item{taxa.col}{String of taxonomic levels to include in classification. Taxonomic levels must exist in the input phyloseq object. If no input given, defaults to all levels in phyloseq object.}

\item{remove}{TRUE or FALSE. Are there characters that need to be removed after the spID in the taxa rowname to isolate the spID ? If TRUE, spID is extracted (e.g sp1_bacteria_bacteroidetes extracted to sp1)}

\item{remove.delimiter}{Character string that separates sp ID from remaining OTU name characters (e.g sp1_bacteria_bacteroidetes delimiter is "_")}
}
\value{
Returns the phyloseq object with taxonomic classification as part of the taxa name (i.e. phyloseq OTU table rownames now contain taxonomy).
}
\description{
Add full taxonomic classification to species ID auto-generated by Phyloseq
}
\examples{
sp_taxa_ID(dat) #Add full taxonomic label to rownames (i.e. sp###_Kingdom_Phylum_Class_Order_Family_Genus)
sp_taxa_ID(physeq = dat, taxa.col = c("Family, "Genus)) #Add full taxonomic label to rownames (i.e. sp###_Family_Genus)
}
